Refactor: Own file for ScanHeader
